P2P: Sanity check for presence of SSID in p2p scan request.
In Change-Id I7d810255a70e5d7e9ae143e094c878f08b2a829d, we
are increasing the dwell time of p2p scan if the p2p scan
is for specific SSID.
There may be chances that some supplicants doesn't trigger
p2p scan for specific SSID or SSIDList is NULL.
This results in kernel panic when p2p scan is issued to
establish p2p connection.

To resolve this issue, the number of SSIDs in scan request
is checked for the presence of SSID in p2p scan request.

Change-Id: I53c7f8a3124ffb9d72301ab2a8c64fade62b871e
CRs-fixed: 591975