CLD: Add iwpriv cmds to set channel time quota & latency for MCC
To help with CST testing for MCC, 2 new iwpriv commands are added to allow
run-time configuration of both channel time and channel latency for home
channels in MCC mode. The commands are:
- iwpriv <MCC mode> setMccLatency <latency in ms>
- iwpriv <MCC mode> setMccQuota <quota in ms>

Notes:
- MCC mode: wlan0 or p2p0. Only MCC modes STA and P2P are currently
assumed.
- latency: minimum non-zero value is 30ms. Otherwise, use 0 for default
setting
- quota: minimum is 20ms, and maximum is 80ms.
- Setting the MCC latency and/or quota for home channels is only possible
iff MCC Adaptive Scheduler feature is disabled.
- Based on the adapter specified in the command, i.e., wlan0 vs. p2p0,
host driver finds out the operating channel and formats both the channel
number and latency (or quota) in a bit vector then passes it to WMA.
- In the case of iwpriv setMccQuota cmd, host driver finds out the other
MCC mode/adapter and its corresponding channel number. Then it formats
channel numbers of both MCC modes/adapters and the time quota specified
in the iwpriv cmd as a bit vector and passes it on to WMI.
- setMccQuota specifies the time quota for a single adapter/mode, while
host driver calculates quota for the 2nd MCC adapter as:
100 - quota_of_first_adapter

Change-Id: If9caec2bb497ffb21c2497de9b448a3095dcea6b
CRs-fixed: 592965