qcacld : Invalid calculation of free wow filters
unsigned interger is assigned -1 value which is causing problem to
calculate the available free filters. Changing the pattern size
as per firmware max size. There is no need to check the number of
allocated patterns in UMAC as this is a redundant, and being taken
care in HDD.Also added one debug print.

Change-Id: Idafbb7e413ddcd1e7dda48c8e298e646757a6792
CRs-fixed: 601456