qcacld: wma: Send vdev restart only during channel switch
VDEV restart was sent to FW during normal connection
as well as during on channel change. Due to this disconnection
is happening if the new channel from csa is not a dfs channel.
With this fix, following sequence is used VDEV start followed
VDEV UP during normal connection and on channel change VDEV
restart followed by VDEV UP on restart resp handler.

CRs-Fixed: 581476

Depends-on: 588153 597667 598919

Change-Id: I5cccf23fdf7a32a1e0a72b6eb2d05af065c21163