qcacld: Implementation of scan timeout on the host driver
When the FW does not send response to the scan command,
the host will be in wait state and will not process any
further commands from the userspace. To avoid this,
timeout is implemented for the scan command on the host
driver. This change is done on the premise that WMA
does not receive any further scan command from the UMAC
until the previous scan command is processed back to the
userspace.

CRs-Fixed: 605502
Depends-on: 598919 588153

Change-Id: I1885fa0f64211b589a2a40f1d858a93235cb7bb2