wlan: copy measurement type and duration (RRM).
Fix for issue where DUT was doing a passive scan
instead of an active scan when there is a beacon report
request on a specific channel. In the case where the
num_APChannelReport is zero, the measurement type and
measurement duration are not being copied. Hence added
an else condition to address this.

CRs-Fixed: 594614