qcacld: Define open source flag based on license
Kbuild file enables WLAN_OPEN_SOURCE flag based on the directory
path but this doesn't work for all the platform as path may be
different. This change passes the WLAN_OPEN_SOURCE flag from
Android.mk or Makefile.  Makefile determines whether the driver
is open source or proprietary and accordingly WLAN_OPEN_SOURCE is
set to 1 or 0.

Change-Id: Id0149f71ceb3e94bd1a67868565cd3b8151bdd7d