qcacld: check for sus_res_mcastbcast_filter_valid before resetting
Check for the value of sus_res_mcastbcast_filter_valid before resetting it
to false. The driver can unload/load while the wlan driver has been
suspended. In such a case, the value of flag sus_res_mcastbcast_filter
would be reset. Check for the value of the flag to take care of this case.

Change-Id: I64ea6f304d4f7535dc1fb47ca3cb0b87654a08ca
CRs-Fixed: 615781