qcacld: Fix link monitor and keep alive period
As of now, the driver is not honoring the link monitor timeout which
is meant for detecting link idleness. And also the host keep alive
period is meant for sending keep alive frame to the associated STA
when it exceeds the maximum idle inactive time. So it has to be
configured as maximum unresponsive timeout in firmware.

Change-Id: I9594e25f19daf4a16de18b4893426e946785c8e7
CRs-Fixed: 613883