qcacld: hdd: Add netlink interface for service
On non-android platform where there is no framework, SSR and
LTE-coex features are not working because of not able to restart
the SoftAP. Adding firmware crashed and LTE Coex indications to
user applications using existing netlink interface service. This
can be extended to other indications required for UI
notifications or other purpose.

Change-Id: If81bf15d16a18427dc55fa4b7a789b8d580b722d
CRs-fixed: 621297