CLD: Correct the usage of gFwWlanFeatCaps
Currently  the capability bits in the gFwWlanFeatCaps are getting set
and retrieved wrongly.
Correcting the usage of this map by left shifting 1 by the number of
positions mentioned by the input capability enum value and then doing
bitwise AND for retrieving the value and doing bitwise OR for setting
capability value.

Change-Id: If75aaad9155ee0859e82045c714aa71b88046f09
CRs-Fixed: 619801