wlan: adding wext interface to set station's mcast rate
HDD received the p2p0 adapter when using wpa_cli to set mcast
rate in wlan0 in station mode. Internal discuss shows that getting
incorrect adapter is an known issue for KK. Due to this, a wext
interface is added as the workaround to use iwpriv to set station's
mcast rate.

Change-Id: Id879507fa217a4467c85002dbdb222a0b5f29284
CRs-Fixed: 622161