qcacld: Fix memory leak in peer->tids_rx_reorder[i].array
'array' is allocated in addba handler and is supposed to be freed in
delba handler. There is the case (for example, in SSR) where delba
handler is not called. Because array points to address of 'base' by
default and is reallocated in addba handler later, only free the
memory when the array does not point to base.

Change-Id: I592afe939895bfd558b8b45f5cf35bc7d9340f69
CRs-Fixed: 620529