qcacld: Fix the timeout value passed to vos wakelock API
- Fix the  timeout value that is passed to the vos wake lock
API. It takes parameter in milli seconds. The calling functions
are actually converting milli seconds to jiffies and passing it
which is wrong. So fix this in all calling functions wherever
the vos wakelock timeout APIs are called with wrong timeout.

- Also, add macros to the wakelock duration which were hardcoded
in WMA.

- Cleanup the WMA wow wakeup handler to handle the different
cases in switch-case instead of cluttering if-conditions.

Change-Id: I3c3389dd06dd56e62812ebd6e50d11e0be70e040
CRs-Fixed: 619610