qcacld: Use HTC Ctrl EP Tx credits for comparison
The current way of getting available HTC credits for WMI uses
the total credits for all the endpoints. We are checking the
number the credits available for WMI while sending the WOW
enable command to firmware. This is leading to abort of suspend
as sometimes the number of host credits available is lesser than
expected. To fix that, use the control endpoint Tx credits before
sending the WoW enable WMI command to firmware. Also, change the
function name accordingly to denote the operation performed on
control endpoint.

Change-Id: Icf00147f2224dfaeff5724dc439887901b1d1446
CRs-Fixed: 625530