wlan: logic errors in TCLAS frame generation APIs
When building the prima driver for 64-bit architecture, the compiler
generated a warning for code in PopulateDot11fTCLAS().

In the present code srcIpAddr and DestIpAddr are defined as a tANI_U32
variable so it is typecasted to tANI_U8 for hidding the compilation
errors in 32-bit.

Defined both srcIpAddr and destIpAddr as an array of tANI_U8.

Change-Id: I64e28af12e8bb537e4a521b570e0b494a6f840f7
CRs-Fixed: 625602