qcacld: txrx: Do not iterate over tx_desc list
While going to suspend we iterate over tx_desc list to find
total desc unused in the freepool. We need not do this
as it is time consuming and we have a variable to track
number of descriptors in freepool. Use this variable
to figure out for any used tx descriptors.

Change-Id: Ifb868e184afe052d44ad5b30cf92e9adcf119176
CRs-Fixed: 626182