qcacld: Abort suspend if internal scan is ongoing
If there is an ongoing scan, the driver should not enable WoW or
suspend the target before receiving scan completion event. Otherwise
the target will go into a state where connection would fail with auth
failure as the target does not seem to process auth responses from AP,
even after several re-transmissions.

If the scan is initiated by supplicant or framework, then device would
not start runtime PM suspend as a wakelock is held till scan completes
and results are posted to cfg80211. We also take care of aborting ongoing
scans in cfg80211_suspend. But this may not be enough if the scan was
started within the driver. In such cases, check if there are any scan
requests submitted to the target and abort the suspend. Later when the
scan completion event is received, we allow the device to go into suspend.

Change-Id: I02e586d47c32d5def67348dc22bc6d35909c74b2
CRs-Fixed: 637401