wlan: VHT IEs missing in Reassoc Request after roam
When roaming from one AP that supports Very High Throughput
(802.11ac, VHT) to another, the 802.11ac rates are no longer
used.  The problem is that there are no VHT IEs in the Reassoc
Request that is sent to the AP that is being roamed to.

To fix this issue, the LIM session entry that is used for the new
BSS has the VHT information set correctly.  Also, the routine
that sends out the Reassoc Request is modified to look at the
VHT information and add the VHT IEs to the frame if appropriate.

Change-Id: I489b3e4b4644f2da23f564e8378213d45a6dd77f
CRs-Fixed:  573910