wlan: qcacld: wma: Fix unused-function warnings
When GCC 4.7 was first introduced to Android, the initial version of
the compiler was warning about things it should not have, and hence
some warnings were suppressed, including unused-function warnings.
The compiler now seems to be correctly warning only about bad
behavior, so clean up the unused functions in WMA in preparation for
re-enabling unused-function warnings in the compiler.

Change-Id: I72fa8e2e51366164845ec8ee5cb41e0985f5be90
CRs-fixed: 640842