qcacld: Fix peer leak in UMAC
It was seen in the code that when DUT is operating in SAP mode,
there is a possibility of peer leak in UMAC, where UMAC does not
do a proper clean-up. This happens when a STA which is trying to
associate to the SAP sends a disassoc when the SAP is waiting for
ASSOC confirmation. UMAC is not doing a DelSTA when processing the
disassoc from the STA. Hence now added a change which will ensure
proper clean-up is done upon receiving a disassoc from the STA.

CRs-Fixed: 642577

Change-Id: I83d32ef60c875c3249c42a501de33d6f1ce323ba