wlan: Update SMD and SMSM header includes (revised)
Commit "wlan: Update SMD and SMSM header includes" added a mechanism
to determine which SMD and SMSM include files are present in the
kernel.  This change utilized modifications to both Android.mk and
Kbuild to make this determination.  This works fine when the wlan
driver is being built as part of an Android build (and hence
Android.mk and Kbuild are both used), but this does not work when the
wlan driver is being built as part of a kernel build (and hence Kbuild
alone is used), such as with Nexus products.  Remove the logic that
was added to Android.mk, and instead add equivalent logic to Kbuild.

Change-Id: I84915b26d1b4ed0b5e1342c38f600132375d78b0
CRs-fixed: 640743