wlan: Flush/filter temp scan results along with the main results
Flush the temp scan results along with main scan results in
csrScanFlushResult(). This is required to flush the results
completely, as the results are getting updated from temp list to
the main list only after completion of scan.
Filter the temp scan results based on valid channels list along
with the main scan results in csrScanFilterResults().

Change-Id: I927948bb4b8b3be402ab67563587be8058ab2a35
CRs-Fixed: 649420