qcacld: Enabled 4ADDR Scheme Flag for STA Mode
Enable 4ADDR WiFi Flag for STA Mode. By default, the STA is brought
up with 4addr disable, which sets IFF_DONT_BRIDGE flag. Due to this
flag we cannot add STA interface as a linux bridge port. This flag
is needed to be set add STA interface as a port to bridge.

Change-Id: I9cb0cd42387ea254c0bfe69f895e2b4b7059810d
CRs-fixed: 642719