qcacld : Fix for crash in gtk offload during wlan suspend
There is a corner case where cfg80211 suspend is successful and
gtk offload configured in f/w but PCIe suspend is fails as scan
starts on some interface. During resume gtk offload is disabled
but as the device is not associated, this step is skipped. When
next time suspend is issued and offloads are configured the f/w
crashes because of double enable of gtkoffload.

Change-Id: I4a86c3a8af8dfdfc7d59939ee4186fddd23ecf82
CRs-fixed: 654775