qcacld: Thermal throttling for high latency device.
When chip temperature is up to upper value which host set to target,
target reports a thermal event to host and host kicks off a phase
timer to throttle traffic. In throttle timer, it has to specify LL
or HL device and then go through different path to throttle. For HL
device, pdev/vdev pause/unpause are used for throttling.

Change-Id: Ic4c3e958a75b5b4c699b716af4f9f5fd987324c2
CRs-Fixed: 665014