WLAN: Don't disconnect if station is connected in supported band.
If station is connected in 2.4 GHz and band capability is changed
from Auto to 2.4 GHz only, station is getting disconnect.
Same is the case with 5 GHz.
Right now disconnection is issued when in connected state
if new band capability is not the supported one.
Disconnect will happen only in case of if connected channel is
not supported in new Band.

Change-Id: I6f42ad5a65dab38eaca1ec23745f8f04e91a882a
CRs-Fixed: 664060