qcacld: Reset the scan info in wma_stop_scan.
In 11R/LFR roaming, scan is started internally by roaming state machine.
The scan info maintained in the wma is not clearing when the scan is
cancelled/completed.

DUT is failing to suspend, because of the scan is pending by reading the
info at wma; which is not updated properly.

Fix by Reseting the scan info at wma in wma_roam_preauth_chan_cancel.

Potential Fix to clear the scan state properly.

CRs-Fixed: 664842
Change-Id: I80a6784687161535433bd184c2c703eba75779eb