qcacld: Clean up suspend/resume code
Cases are seen where PCIe interrupts are not recieved for which
host is not able to clean up all the allocated wmi buffers.

This resulting is fail to suspend by reading pending cmds at wmi
layer, which is reduced in Txcompletion handler.

Other instances are seen where HDD timesout waiting to get the
suspend indication from wma.

The wmi_is_suspend_ready is sleeping for 1.6 sec and if there are
pending commands which are not cleared, then the timeout at HDD;
i.e 2sec, is kicked out before the wma returns.

Since there is enough protection logic at hif_pci_suspend to ensure
the credits avaiable before posting the wow_enable commands and after
receiveing HTC_ACK, the code to check for pending commands is redundant
and slows the system; as we are sleeping the MC thread for almost
1.6sec in the worst case.

In the worst case; we will pause the tx quques for 2sec; even if
suspend fails.

Fix by offloading some work to resume path from suspend path.

Change-Id: I888f6e8af90fdedfa88b3769d04e9aec0b9c9dcc
CRs-Fixed: 664649