qcacld: Fix for possible MIC error while roaming
In recent roaming tests it was seen that roaming failed
sometimes because of MIC mismatch on AP side, reason being
AP sometimes sent a PreAuth again with a modified Anonce
(may be because of ACK miss). But the driver drops the
subsequent preauth frames and calculates the MIC on first
received Anonce whereas AP calculates MIC on last sent Anonce,
and hence a mismatch.
Changes have now been made to update the saved FTIEs in the
driver if there are any subsequent preauth frames, however this is
valid as long as the message is not posted to SME. This would help
avoid MIC errors atleast in most cases where the APs are on different
channels.

Change-Id: I9503906bb3df8bc6e3c8c4aa40c358572207d2a1
CRs-Fixed: 660796