wlan: add compat support to iw_setnone_getnone()
The WLAN_PRIV_SET_NONE_GET_NONE ioctl is a special case where a
sub-ioctl is used and both the number of get and set args is 0.  In
this specific case the logic in iwpriv places the sub_cmd in the
data.flags portion of the iwreq.  Unfortunately the location of this
field will be different between 32-bit and 64-bit userspace, and the
standard compat support in the kernel does not handle this case.  So
we need to explicitly handle it iw_setnone_getnone().

Change-Id: Id08284917ac70a9e93bd90c50b9aa8384eb12845
CRs-fixed: 670292