qcacld: Configure proper max tx power to firmware
When processing beacon in the case of VoWiFi RRM, the max
Tx power calculation is not taking the host Min/Max Power
macros into consideration, because of which host is configuring
a power which is beyond the max Power macro. The necessary changes
to fix this have been made in this gerrit.

Also, in WMA the max tx power is not updated in all the cases,
made changes necessary for the same.

Change-Id: I9749cf779bc8921da477d3d087989133f857c8e0
CRs-Fixed: 671823