qcacld: Optimize burst buration for normal scan in case of p2p
Currently, the burst duration is equal to the max channel time
and hence only one channel will be accomodated in the burst
scan. This is optimized for miracast performance. But, when
miracast is not running, max channels can be accomodated in
burst scan to make the scans faster. Fix that!

Change-Id: I523af13d43102b9d96cc2b167cf5addc30ecde3c
CRs-Fixed: 676484