Add coex for AP scan NULL case
When AP scan result is NULL, it used to mean that there are no interfering
sources, so SAP can operate on the start channel of configured channel
ranges. This is not the case when there are internal interference, from
cellular modem. This change in the channel selection process addresses
the issue.

Change-Id: Id93cd28dd6550c8c2634d43103e4f8b7a6391704
CRs-Fixed: 672242