qcacld: Fix of HDD getting stuck in disconnecting state
HDD is stuck in disconnecting state because by the time
HDD calls disconnect API SME is already in disconnected
state so it does not respond to HDD request and after it
HDD is stuck in disconnecting state. Fix is to change
connState to disconnected state in wlan_hdd_disconnect
after completing wait for completion. In addition added
debug code to check if HDD is getting stuck because of
SME disconnect API returns failure.

Change-Id: I383414833c173dc2783918a7df434fd6a4d8bb2f
CRs-Fixed: 677107