qcacld: HDD_VOSS: Restrict creation of multi-port concurrency
Checks are added to restrict the creation of multi-port concurrency.
As of now, only two-port concurrency is supported; if there exists
a 2-port concurrency then the third connection is not allowed.
However, it does not restrict the creation of interface.

For example:
1. STA + P2P (GO/CLI) connection exists and then new
   connection establishment is not allowed.
2. STA + SAP connection exists and then new
   connection establishment is not allowed.
3. P2P GO + P2P CLI connection exists and then new
   connection establishment is not allowed.
4. STA + ADHOC connection exists and then new
   connection establishment is not allowed.

In the above example combinations, if any of the existing connection
is disconnected, then next connection is allowed.

Change-Id: Ia49994409b90b2eacbdb9dfa5bfa407cd089a5de
CRs-Fixed: 668296