wlan: Cannot change PMF SA Query configuration
The PMF SA Query configuration parameters "maximum
retries" and "retry interval" cannot be changed by
setting to new values in the WCNSS_qcom_cfg.ini
file.

The problem is in the definition of these parameters
in the configuration database.  They were defined as
read-only, when they should have been defined as
read-write.

Because of this issue, the driver always used the
default values for the parameters instead of the
new values.

Change-Id: I99307af9ca8843e24c65d1f9d2035af6ddeb2de0
CRs-Fixed: 679585