qcacld: HDD: Provide fix for passing right concurrency combinations
cfg80211 expects the driver to advertize the supported intf and
concurrency combinations. This combination will be used by the
kernel(read cfg80211) to restrict the user from creating unexpected,
unsupported concurrency scenarios and also restrict exceeding maximum
supported interfaces in a given device mode(STA/P2P/SAP etc).
The experiments have shown that cfg80211 will let the interface to be
created but keeps the interface busy so that if won't be "up"ed to
start any new connection.

Change-Id: I21e355d1229b58aca29455513dfa68d3ffead45c
CRs-Fixed: 668290