qcacld: hdd: send protocol reason code for auth fail
When attempt of connection fails, driver sends the generic
reason code instead of sending protocol reason code.

This restricts application not to reconfigure station to
SHARED-WEP when OPEN-WEP authentication fails.

This change has been done in a way that driver will send exact
protocol reason code in case of OPEN-WEP or SHARED-WEP besides
station interface is not in connected state.

Change-Id: I33fb59a3654fbfbf113376a737ab4d60bcc12749
CRs-fixed: 679890