wlan: Dont send deauth for PMF sta,if auth received in connected state
In SAP if connection is PMF enabled and auth is received in connected
state from the station do not send deauth.

For PMF connection the AP should not tear down or otherwise modify
the state of the existing association until the SA-Query procedure
determines that the original SA is invalid.

Change-Id: Ib9ba7981e6b54fce1ea072dafe5dfe4cb0d0ca72
CRs-Fixed: 680218