qcacld: wma: fix memleak
In case of SAP running in DFS channel, ic_curchan would be allocated
while it was first brought up. When it detected the radar and
restarted the SAP after randomly picking channel, the ic_curchan was
still holding the memory allocated. Add sanity checking and release the
memory if it was allocated.

Change-Id: Ie9f92c636ad0e15444403ae2a5ed670eca23c554
CRs-fixed: 680517