qcacld: sap: add iwpriv command to send fake radar signal event
In order to sanitize the post radar detection logic without real radar
injection equipment, this command is added to self indicate a fake radar
event, to trigger the post radar detection logic, including the CSA and
channel switch etc.

The change is to bypass the real radar pattern matching and directly
call to the wmi interface to indicate HDD and SAP to stop the traffic
and to stop beaconing, and then start the channel selection and send out
CSA to switch the channel.

Change-Id: Ic6fc782a31a2fba078be1b427cb85e0cbfeaa25a
CRs-fixed: 671191