qcacld: ExtScan: Fix incorrect parsing of MAC address
The problem is because of the usage of nla_memcpy along with nla_data.
The 2nd argument pointer is advanced by NLA_HDRLEN using nla_data, but the
implementation of nla_memcpy takes care of advancing the pointer
by NLA_HDRLEN using nla_data() internally.
This caused the pointer to advance two times and end up reading the
data 4 bytes ahead of where it was supposed to read.

Fix this issue by not using nla_data() for the source attribute and let
nla_memcpy take care of this

Change-Id: I935dacc033ffc1100f799b88212feec5db72c152
CRs-Fixed: 686852