wlan: scan taking more than 1 sec despite of dwell time values.
As part of scanning, we step through the list of channels and
calculate the "total scan duration" and initialize the HAL
with this value. The problem here is that we use the "dwelltime"
values set through the ioctl interface only if the  channel is
marked for "Active scan". For DFS channels that are marked for
passive scan, we retrieve the values from cfg and the default
value for it is 110msec. Since we have so many DFS channels
in the 5GHZ region, the final total scan duration is coming
to be much more than 1sec. The fix is to update the cfg with
the appropriate dwell times values when they got set through
the ioctl interface.

Change-Id: I8ea59f4f87d966f7a76cb41879ed7ae3e132b95d
CRs-Fixed: 696844