wlan: hdd: Remove obsolete ISOC and WIFI_2_0 conditions
The qcacld driver contains a large amount of conditional compilation
that was initially added to allow the driver to support both
integrated and discrete wlan solutions.  However it has been decided
that the integrated solutions will only be supported by the prima
driver, so remove from HDD all of the conditional compilation that is
now obsolete.

- Code that is conditional upon QCA_WIFI_ISOC being defined is
  removed.
- Code that is conditional upon QCA_WIFI_ISOC not being defined is
  retained without conditional compilation.
- Code that is conditional upon QCA_WIFI_2_0 being defined is
  retained without conditional compilation.
- Code that is conditional upon QCA_WIFI_2_0 not being defined is
  removed.

Change-Id: I5d11c5d7c4bb5d45ca98c491633a26d1be69e29a
CRs-fixed: 698342