qcacld: sap: add channel matrix info for DFS channel selection
In DFS testing, when switching to other channel, there are chances
are some channel will have tx leakage across the band, which would
break the certification testing.

The table of channel matrix information is added with the tx leakage
level according to the source and target channels.

Add the logic to check the channel condition to avoid the channels
that are above certain tx leakage threshold.

Change-Id: I223d5cdacaaff984c3fdea3a4cd775ce86194119
CRs-fixed: 698338