qcacld: ipa: Dont request RM when no pending packets
With the recent changes to prevent the suspend while data is
going over IPA, a new issue is introduced where in IPA RM request
would be called even when there are no nbufs in the pending
queue.  This leads to not releasing the RM as release happens
from write done call back which would never be called as no nbufs
are passed to IPA hardware. This issue happens when all the nbufs
are pre-filtered and no nbufs are queued to pending queue to be
sent to IPA.

Also along with this change, adding debugfs write call back to
reset the stats which would help for better debugging and
increase the debugfs buffer size to 4096 to support concurrent
interfaces. Buffer size of 2048 is not enough to hold all the
interface statistics.

Change-Id: Iaa876e2651cdc3400fb9a4de9f85ac7a7a5e05af
CRs-fixed: 698929