wlan: qcacld: Remove pronto support from Android.mk
It has been decided that the qcacld driver will no longer support the
pronto wlan core, so we want to remove all conditional compilation
that exists to support this target.  Start by removing the conditional
logic that is used to pass WLAN_ISOC_SELECT from Android.mk to Kbuild.

Change-Id: I35991068424d7c10be2e42c2af458c75955c5207
CRs-fixed: 699601