wlan: sme: Remove obsolete ISOC and WIFI_2_0 conditions
The qcacld driver contains a large amount of conditional compilation
that was initially added to allow the driver to support both
integrated and discrete wlan solutions.  However it has been decided
that the integrated solutions will only be supported by the prima
driver, so remove from SME all of the conditional compilation that is
now obsolete.
- Code that is conditional upon QCA_WIFI_ISOC being defined is
  removed.
- Code that is conditional upon QCA_WIFI_ISOC not being defined is
  retained without conditional compilation.
- Code that is conditional upon QCA_WIFI_2_0 being defined is
  retained without conditional compilation.
- Code that is conditional upon QCA_WIFI_2_0 not being defined is
  removed.

Change-Id: Ia19bc22c1fcb84799f15d5297637c4f065489d58
CRs-fixed: 702843