wlan: wda: Remove obsolete ISOC and WIFI_2_0 conditions
The qcacld driver contains a large amount of conditional compilation
that was initially added to allow the driver to support both
integrated and discrete wlan solutions.  However it has been decided
that the integrated solutions will only be supported by the prima
driver, so remove from WDA all of the conditional compilation that is
now obsolete.
- Code that is conditional upon QCA_WIFI_ISOC being defined is
  removed.
- Code that is conditional upon QCA_WIFI_ISOC not being defined is
  retained without conditional compilation.
- Code that is conditional upon QCA_WIFI_2_0 being defined is
  retained without conditional compilation.
- Code that is conditional upon QCA_WIFI_2_0 not being defined is
  removed.

Change-Id: Ie8d0f811e48096b906acb5dce15806364a42c8f1
CRs-fixed: 702861