qcacld: Fix the probe repeat time in concurrency scenario
Currently in FW, during the probe repeat time it will
send 3 probe requests most of the time and in the host,
in STA+SAP concurrency, the dwell time value is limited
to 28ms.

In such a scenario with the current implementation,
the probe repeat time is not sufficient to address the
given number of SSIDs in the scan request. This fix will
add a static table which will decide the probe repeat
time based on the number of SSIDs in the scan request.

Change-Id: Ieed07cc45f0a60f7c4a90389e95e566a8fa26d76
CRs-Fixed: 704690