wlan: tl: Remove obsolete WIFI_2_0 conditions
The qcacld driver contains a large amount of conditional compilation
that was initially added to allow the driver to support both
integrated and discrete wlan solutions.  However it has been decided
that the integrated solutions will only be supported by the prima
driver, so remove from TL all of the conditional compilation that is
now obsolete.
- Code that is conditional upon QCA_WIFI_2_0 being defined is
  retained without conditional compilation.
- Code that is conditional upon QCA_WIFI_2_0 not being defined is
  removed.

Change-Id: I4e6fb1b1068add638cb386ef1aef5ef55548ae11
CRs-fixed: 709179