wlan: Kbuild: Remove obsolete WIFI_2_0 conditions
The qcacld driver contains a large amount of conditional compilation
that was initially added to allow the driver to support both
integrated and discrete wlan solutions.  However it has been decided
that the integrated solutions will only be supported by the prima
driver, so remove from Kbuild all of the WIFI_2_0 conditional logic
that is now obsolete.

Change-Id: I35beb3e01278eb91f9eb5e56e3c167e06bca3a98
CRs-fixed: 709179